requirements.txt
jsonpatch==1.33
langchain==1.1.3
langchain-core==1.1.3
langchain-ollama==1.0.0
ollama==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
streamlit==1.52.1
tenacity==9.1.2
tornado==6.5.2
typing_extensions==4.15.0
wikipedia==1.4.0
